backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.35
alembic==1.13.2
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2
markdown==3.7
